Fixed slideshow no base editor error 2
There was an error when if you had no scripts open at all and the slideshow was running, it'd give an error saying it couldn't get the base_editor on a null value in the clear folder theme function. But if you opened a new script it would have a different BG (showing the slideshow worked). So this change got rid of the error.